Remove alias and fix cache table SQL
